# Enable unicode support in pcre