--- a/src/core/regex.cc
+++ b/src/core/regex.cc
@@ -48,9 +48,6 @@ Regex::Regex(const char16_t *pattern, uint32_t pattern_length, u16string *error_
 
   u16string final_pattern = preprocess_pattern(pattern, pattern_length);
 
-  pcre2_compile_context *context = pcre2_compile_context_create(nullptr);
-  pcre2_set_newline(context, PCRE2_NEWLINE_ANY);
-
   int error_number = 0;
   size_t error_offset = 0;
   uint32_t options = PCRE2_MULTILINE | PCRE2_UTF;
@@ -61,9 +58,8 @@ Regex::Regex(const char16_t *pattern, uint32_t pattern_length, u16string *error_
     options,
     &error_number,
     &error_offset,
-    context
+    nullptr
   );
-  pcre2_compile_context_free(context);
 
   if (!code) {
     uint16_t message_buffer[256];