--- a/src/core/regex.cc
+++ b/src/core/regex.cc
@@ -40,7 +40,7 @@ static u16string preprocess_pattern(const char16_t *pattern, uint32_t length) {
 }
 
 
-Regex::Regex(const char16_t *pattern, uint32_t pattern_length, u16string *error_message, bool ignore_case) {
+Regex::Regex(const char16_t *pattern, uint32_t pattern_length, u16string *error_message, bool ignore_case, bool unicode) {
   if (pattern_length == 0) {
     pattern = EMPTY_PATTERN;
     pattern_length = 4;
@@ -50,8 +50,9 @@ Regex::Regex(const char16_t *pattern, uint32_t pattern_length, u16string *error_
 
   int error_number = 0;
   size_t error_offset = 0;
-  uint32_t options = PCRE2_MULTILINE | PCRE2_UTF;
+  uint32_t options = PCRE2_MULTILINE;
   if (ignore_case) options |= PCRE2_CASELESS;
+  if (unicode) options |= PCRE2_UTF;
   code = pcre2_compile(
     reinterpret_cast<const uint16_t *>(final_pattern.data()),
     final_pattern.size(),
@@ -74,8 +75,8 @@ Regex::Regex(const char16_t *pattern, uint32_t pattern_length, u16string *error_
   );
 }
 
-Regex::Regex(const u16string &pattern, u16string *error_message, bool ignore_case)
-  : Regex(pattern.data(), pattern.size(), error_message, ignore_case) {}
+Regex::Regex(const u16string &pattern, u16string *error_message, bool ignore_case, bool unicode)
+  : Regex(pattern.data(), pattern.size(), error_message, ignore_case, unicode) {}
 
 Regex::Regex(Regex &&other) : code{other.code} {
   other.code = nullptr;