… vendor/pcre/include/config.h