Enable Unicode support in pcre and compile regular expressions with `PCRE2_UTF`.

Fixes #56.