@smashwilson 2018-03-28
This does have a not insignificant performance and size impact, hence the failing test:

##### Without `SUPPORT_UNICODE`

* Size of `superstring.node`: 768K
* Duration of `"can be called repeatedly between buffer mutations without harming performance"`: 121ms
* Size of `browser.js`: 808K
* Duration of `"can be called repeatedly between buffer mutations without harming performance"` with browser tests: 715ms

##### With `SUPPORT_UNICODE`

* Size of `superstring.node`: 896K
* Duration of `"can be called repeatedly between buffer mutations without harming performance"`: 440ms
* Size of `browser.js`: 1.1M
* Duration of `"can be called repeatedly between buffer mutations without harming performance"` with browser tests: 3059ms

_All recorded on macOS, naturally._

I'm not sure what to do with this one from here. That's a pretty big hit and this is a performance-critical part of Atom, but not handling unicode surrogates and character attributes at all is bad for internationalization.

One idea I have is to remember which `Text` instances contain surrogate pairs and which do not, and only compile a `Regex` with `PCRE_UTF` if it will actually be necessary. That would at least limit the penalty to cases where it's actually necessary... although if this applies to a user it likely apples to most of the buffers they open.

Another option might be to handle this during pattern preprocessing: replace `á` with `[áÁ]`, for example. I suspect the end game there is us reimplementing what `PCRE_UTF` does, but poorly.

@nathansobo @maxbrunsfeld: What do you think is the right thing to do here? Is this why `SUPPORT_UNICODE` was off in the first place?