@maxbrunsfeld 2018-03-28
@smashwilson I don't recall intentionally turning unicode support off; I think I just wasn't aware of (and didn't realize that we weren't handling) regexes with `/u`.

I'm assuming that the performance degradation only occurs when the `PCRE2_UTF` flag is set? If that's the case, then can we only set that flag when the JavaScript regex has the `/u` flag and contains non-ascii characters?