@smashwilson 2018-03-28
> I'm assuming that the performance degradation only occurs when the PCRE2_UTF flag is set?

Verifying that presently.

> If that's the case, then can we only set that flag when the JavaScript regex has the /u flag and contains non-ascii characters?

Ah! I didn't know JavaScript regexes had a `/u` flag. Okay, I'll go that route. I might put the "detect when to add `/u` to the regex" logic in find-and-replace then.