@maxbrunsfeld 2018-03-28
Actually it's a bit more complicated than detecting non-ascii characters. It seems like `/u` affects the behavior of negated character classes and the `.` operator, and adds a `\u` syntax.

https://mathiasbynens.be/notes/es6-unicode-regex