@smashwilson 2018-03-28
> I'm assuming that the performance degradation only occurs when the PCRE2_UTF flag is set?

Looks good: about 100ms with node, 1000ms in browser.

> https://mathiasbynens.be/notes/es6-unicode-regex

I just found that too :smile:

Well that's a bit of a mess. I'd suggest adding another toggle to find-and-replace to flip `/u` but that would just confuse everyone.

Of the `/u` changes I'd guess that we would generally want the new `.` behavior and the `\u` syntax. The negated character classes and some of the wrinkles of canonicalization and case folding feel riskier, but they at least seem to be the same thing that pcre2 does: https://www.pcre.org/current/doc/html/pcre2unicode.html.