@smashwilson 2018-03-28
Offhand I'm leaning toward "set `/u` on find-and-replace regexes (and therefore pcre2 regexes)" when the buffer text _or_ the pattern source contain astral-plane characters or `\u{..}` sequences. That seems like a reasonable rule of thumb, and tracking when a buffer contains or does not contain astral-plane characters should be doable as part of any other O(n) scan we have to do on the content anyway.