@smashwilson 2018-03-29
> tracking when a buffer contains or does not contain astral-plane characters

I'm punting this to a separate PR because "respecting `/u`" stands alone well and because I'm much less confident about the implementation there :smile: